requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
djangorestframework==3.14.0
drf-url-filters==0.5.1
Pillow==9.2.0
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
voluptuous==0.13.1
